+++ tests/ttnn/unit_tests/operations/eltwise/test_unary.py
