+++ ttnn/cpp/ttnn/operations/eltwise/unary/unary_pybind.cpp
@@ -1714,7 +1714,7 @@ void py_module(py::module& module) {
         ttnn::trunc,
         R"doc(\mathrm{{output\_tensor}}_i = \verb|trunc|(\mathrm{{input\_tensor}}_i))doc",
         "",
-        R"doc(BFLOAT16, BFLOAT8_B)doc");
+        R"doc(FLOAT32, BFLOAT16, BFLOAT8_B)doc");
     bind_unary_operation(
         module,
         ttnn::eqz,
